requirements.txt
# Dépendances pour le module d'ingestion OpenMetadata Dremio
# Installation avec venv: 
#   python -m venv venv
#   venv\Scripts\activate (Windows) ou source venv/bin/activate (Linux/Mac)
#   pip install -r requirements.txt

# ============================================================================
# CORE DEPENDENCIES - API et HTTP
# ============================================================================

# HTTP client pour API calls OpenMetadata
requests>=2.31.0

# YAML parser pour fichiers de configuration d'ingestion
PyYAML>=6.0.1

# URL utilities (dépendance de requests, version compatible)
urllib3>=2.0.0

# JSON schema validation (pour validation des configs OpenMetadata)
jsonschema>=4.19.0

# ============================================================================
# DATABASE CONNECTORS
# ============================================================================

# PostgreSQL connector (pour Dremio via JDBC PostgreSQL)
psycopg2-binary>=2.9.7

# OpenSearch connector (pour ingestion directe)
opensearch-py>=2.4.0

# ============================================================================
# DATA PROCESSING
# ============================================================================

# Date/time utilities
python-dateutil>=2.8.2

# Typing extensions pour compatibilité Python
typing-extensions>=4.7.0

# Charset detection (améliore requests)
charset-normalizer>=3.2.0

# Certificate management
certifi>=2023.7.22

# ============================================================================
# LOGGING ET MONITORING  
# ============================================================================

# Enhanced logging avec couleurs
colorlog>=6.7.0

# Structured logging
structlog>=23.1.0

# ============================================================================
# OPTIONNEL - OPENMETADATA INGESTION (si besoin framework complet)
# ============================================================================

# OpenMetadata Ingestion Framework (commenté par défaut - très lourd)
# openmetadata-ingestion[postgres,opensearch]==1.9.7

# ============================================================================
# DEVELOPMENT (optionnel - décommentez si besoin)
# ============================================================================

# Testing framework
# pytest>=7.4.0
# pytest-cov>=4.1.0

# Code formatting
# black>=23.7.0
# flake8>=6.0.0

# Type checking  
# mypy>=1.5.0

# ============================================================================
# NOTES D'INSTALLATION
# ============================================================================
# 1. Ce fichier est optimisé pour un module léger d'ingestion
# 2. Évite les dépendances lourdes comme Airflow sauf si nécessaire  
# 3. Compatible Python 3.8+
# 4. Testé avec OpenMetadata 1.2+ et 1.3+